- Split pseudo into VGSavePseudo & VGRestorePseudo.
- Check for noexcept or nothrow when emitting Clang errors.
